input-leap: update port
Update to latest git HEAD.

Include a patch to remove arbitrary mouse wheel scroll speed multiplier.
This can be removed after scroll speed on macOS is properly addressed
upstream.

Add a note for dealing with application permissions during upgrade.
This, too, is a stop-gap measure and not an ideal solution.