fnm: fix fish completions location
`$destroot/$prefix/share/fish/completions` is for the completion files
included with fish, not for use by package
managers installing additional files. The correct location is
`$destroot/$prefix/share/fish/vendor_completions.d`.

This change leaves the completion generated commented out, but makes it
point to the correct place if the decision is made to reintroduce this
(I think it should be reintroduced, as is done with `ripgrep/Portfile`).