openssl3: Fix CVE-2023-5678
Backport a fix for a low severity denial of service vulnerability in
X9.42 DH. See

  https://mta.openssl.org/pipermail/openssl-announce/2023-November/000284.html

for the upstream advisory. You are very likely not affected by this,
because X9.42 DH keys are not very common anymore. If you are, now would
be a good time to move to well-known safe prime groups for
Diffie-Hellman.

No revbump of the openssl port, since this does not change which files
are installed. Also no revbump of any of the dependent ports, since this
fix does not change ABI.

CVE: CVE-2023-5678