libgrapheme: new Portfile
* new Portfile

libgrapheme is "an extremely simple freestanding C99 library providing
utilities for properly handling strings" from suckless.org.

As with many suckless.org projects, build settings are configured via
config.mk. There is a configure script, which simply sets values in
config.mk based on the OS. config.diff applies the Darwin build
settings and removes the hardcoded PREFIX from config.mk.

libgrapheme is a dependency of another suckless.org project, lchat,
which is a frontend for their ii IRC client. If this port is approved,
I hope to add a port for lchat.