lima: fix missing version from limactl info
The newest release of colima now checks for the version of lima, but the
lima Makefile expects to build from a checked out git file, not
a tarball. It sets `VERSION` from the latest git tag. This patches the
Makefile to replace the `git` version check and replaces it with the
version determined from the Portfile.