watchman, edencommon: Update to 2023.11.13.00
Fix packaging of pywatchman and specify the watchman version number
(which cannot be determined from git in our setup).

Drop the first patch, which does no longer seem to be required.

Update edencommon to the same version, because the build will fail
otherwise.

This fixes the watchman prebuilt package, which is broken since the last
libsodium update in 86381201edd2b233a20f15a542eae461de5eabb6 and was not
fixed in https://github.com/macports/macports-ports/pull/20537 either.

Closes: https://trac.macports.org/ticket/65475
Closes: https://trac.macports.org/ticket/68702