stack: force only system libs
because macports' libiconv is free in the ${prefix}, it is
picked up opportunistically by the build of ghc.

macports version of libiconv is different than the system version
of libiconv. The bootstrap ghc compiler is built against the
system libiconv, and so requires linking against that version.

during the build of stack, pkg-config calls for the zlib.pc. In
macports, the zlib.pc file brings in flags that link against the
libraries in ${prefix}/lib, where libiconv is then found.

to get around this, bring in a modified zlib.pc that uses only
system paths. This forces the use of and link against the system
zlib, and prevents libiconv in the ${prefix} from being
opportunistically linked.