gcc7+: enforce using system `as`
Since GCC 7.4.0, during configure, it detects features supported by target-as.
On the other hand, MacPorts cctools contain a proxy for `as` that runs system
`as` or one of the supported MacPorts clang's `as` if it is installed. Here,
we may encounter a misconfiguration when GCC builds on a system with some
MacPorts clang, and `as` is using it. However, on a clean system, it uses
system `as` if no MacPorts clang is installed, which may behave differently.
This can make GCC broken until MacPorts clang is installed. To avoid a stealth
dependency on the used clang, I enforce building with system `as`.

See: https://github.com/gcc-mirror/gcc/commit/b410cf1dc056aab195c5408871ffca932df8a78a
Closes: https://trac.macports.org/ticket/68683

[skip ci]