moar: Add +pager variant to install as moar-pager
This implements the alternative route described in trac:68491 to enable
a compatibility workaround to use both textproc/moar and lang/MoarVM
since they both wish to install `moar` binaries.

Previously, #20930 marked moar and MoarVM as mutually conflicting. This
conflict state remains, but a new `+pager` variant is added which
installs `textproc/moar` as `moar-pager`. The manpage is also installed
as `moar-pager.1` and the contents are patched to reflect the updated
name.

This variant *removes* the `MoarVM` conflict marker.

Closes: https://trac.macports.org/ticket/68491