OpenBLAS: fix build on 10.6 x86_64
`CMAKE_SYSTEM_PROCESSOR` is output of `uname -m` which might not be the
desired userland architecture. On macOS, it's the kernel architecture,
and on some x86_64 machines running Mac OS X 10.6 the kernel is 32-bit,
which is of course not a problem at all for running 64-bit userland
programs.

Anyway, on that case the build will turn to 32bit on that machine and
fail.