boost176: allow univeral building on arm/Intel
boost176 predates the arm Macs by some years

We use a fallback pathway to force the "combined" architecture
and jam the proper archflags into the build via that method.

Then there is an issue properly selecting a set of asm sources
to use. The hiccup is on the abi, which is different between
arm64 and Intel, so we remove the abi from the specifications
for asm files.

With these two changes, boost176 builds through as universal
on arm and has the required architectures in libraries.

This is not elegant, but it works until our default boost is advanced
to some newer version that better supports arm/Intel universal builds.

closes: https://trac.macports.org/ticket/64954