legacy-support: Keep v1.0.13 on 10.4. (#21775)
Versions 1.1.0 and 1.1.1 fail to build on 10.4.  This hack should
be removed once there's a release that works on 10.4.

TESTED:
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.15 x86_64, and
11.x-14.x arm64.
As before, 1.0.13 on 10.4 builds and passes tests in all cases except
ppc +universal, which fails to build.
On 10.5+, the unchanged 1.1.1 still builds in all cases and fails
tests in some.