hdf5: Avoid duplicate LC_RPATH's (#21962)
.

* Avoid duplicate LC_RPATH's in hdf5 fortran dylibs, using new compiler PG option.
* Rev bump to push out the corrected dylibs.

This PR attempts to fix builds for dependent ports on Ventura and Sonoma, with Xcode 15.  The Xcode 15 linker has a new rule against duplicate LC_RPATH's.  Rev bump is needed to push out the corrected dylibs.

The fix for duplicate LC_RPATH's was copied from the recent OpenBLAS fix for the same thing:
https://github.com/macports/macports-ports/pull/21765