qt5-qtcreator: remove license_noconflict
Appears to be distributable without it now. It was incorrect to specify
openssl as non-conflicting, because if it did somehow start using use
openssl there would be a real conflict. You specify the thing that you
do use that is appearing to cause a conflict but actually doesn't, not
something that you don't actually use.