expat: Disable tests for OS X 10.8 and earlier
As of expat 2.6.0, it verifies that the compiler supports C++11 when
the tests are enabled. The tests already required C++11 before, the
configure script just didn't check for that before. This causes a
dependency cycle on pre-10.8 systems that don't ship with a C++11
compiler because MacPorts clangs eventually depend on expat.