legacy-support: Fix universal build on 11.x
On macOS 11, optool is used to construct the LegacySupportSystem
library, but that doesn't work on arm64.  Prior to this fix, the
decision to use it was based on build_arch, which is problematic for
universal builds in two ways:

1) Universal builds on x86_64 didn't work at all, due to attempting to
apply optool to the arm64 slice.

2) Universal builds on arm64 failed to apply optool to the x86_64
slice.

This reworks the logic to base the optool dependency on whether any
non-arm64 slice will be built, and to base the decision to use it
on the architecture of each individual slice.  It does not fix the
problem where LegacySupportSystem is missing on arm64 -universal,
and lacks the arm64 slice on +universal.

In principle, this should get a revbump due to the change in the
installed content in the arm64 +universal case.  But since
legacy-support isn't even applicable to 11.x, that doesn't seem worth
the unnecessary churn.

TESTED:
Now builds successfully on 11.x x86_64 and arm64.
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.4-12.x x86_64, 11.x-14.x
arm64.