legacy-support: Fix test failures on 10.6-10.8
The C++ tests were failing to build on platforms with a non-default
cxx_stdlib selection (10.6-10.8).  Since the "real" code is C-only,
and the stdlib selection is irrelevant to testing the use of the C
library, we simply disable cxx_stdlib.

TESTED:
Tests now build and pass on 10.6-10.8 +/-universal.
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.4-12.x x86_64, 11.x-14.x
arm64.