libffi: Fix build on 10.4, 10.5
This makes three changes to the patches for 10.4 and 10.5:

1) In the G3 patch, it removes the ppc7400 reference that was never
actually used, rather than updating it.  It still fixes the case that
is actually used.

2) It removes the patch for i386 assembler that has been
(approximately) incorporated upstream since v3.4.5.  The collision
with this patch was the reason for the patch failure.

3) For the gcc 4.2 test-build fix, it replaces the conditionally
applied patchfile with an unconditional patch controlled by an
environment variable, thereby keeping the source code consistent
across OS versions.  This new environment variable is now defined by
the Portfile when the compiler is gcc 4.2.

The patches for #1 and #3 are now in a single unified patchfile.

Closes: https://trac.macports.org/ticket/69380

TESTED:
Tested on 10.4-10.5 ppc, 10.5-10.6 ppc (x86_64 Rosetta), 10.4-10.6
i386, 10.4-12.x x86_64, and 11.x-14.x arm64.
Now builds and passes at least some tests on all tested platforms,
except 10.4 ppc +universal.
Test behavior on 10.6+ is essentially unchanged.  Tests on 10.4 and
10.5 have a few additional failures since v3.4.4 (the last version
that built successfully) due to new tests having been added.