libffi: Several Portfile fixes and improvements
This makes a number of changes to the Portfile:

1) Removes the superfluous dependency on expect.  This was probably a
holdover from a time when dejagnu was missing the expect dependency.
Not only is this dependency superfluous (since the tests don't
reference expect directly), but it's also more heavy-handed since
dejagnu's expect dependency only applies when the OS-supplied expect
is too old.

2) Fixes the build to get the default architecture correctly.  This is
a general fix that replaces the narrow fix for the 10.6/ppc/Rosetta
case.  See the Portfile comments for details.

3) Removes the tweak to force the use of the integrated assembler when
building with clang.  This is now obsolete (since all builds work
without it), and causes some slightly inferior results in some tests.

4) Made the clearing of cxx_stdlib unconditional.  This was only
being done in the specific case where not doing so led to a circular
dependency, but since libffi doesn't use C++ at all, it's preferable
to do it in all cases and avoid unnecessary compiler constraints.
See the Portfile comments for more details.

5) Adds myself as the maintainer (previously nomaintainer).

Since #3 may affect the installed content in some cases, it includes a
revbump.  No revbumping of dependents is necessary.

TESTED:
Tested on 10.4-10.5 ppc, 10.5-10.6 ppc (x86_64 Rosetta), 10.4-10.6
i386, 10.4-12.x x86_64, and 11.x-14.x arm64.
Builds on all tested platforms except 10.5 ppc +universal, with
test results equal to or better than the previous state.