gstreamer1-gst-plugins-base: Don't conflict with gl-headers & mesa
When the GL headers are not found meson will download them using  the gl-headers wrap, this causes gstreamer1-gst-plugins-base to conflict with gl-headers port & mesa.

Use the gl-headers Port to avoid embedding the headers, this also avoids the download from failing as it requires a modern version of git to download the subproject.