ruby32: update to v3.2.4; fix build on 10.4, 10.5
See:
https://www.ruby-lang.org/en/news/2024/04/23/ruby-3-2-4-released/

This incorporates the applicable patches from ruby30, including
converting the reinplace patch for the gem program name to a normal
source patch.

It adds two new patches related to the CommonCrypto headers on
10.4 and 10.5.

All source patches are in a single patchfile derived from the ruby
fork at github.com/fhgwright/ruby.  The changes are much more readable
in the git repo.

It also narrows the compiler blacklisting, corresponding to what was
done in ruby30.

It also incorporates the cstdbool fix for 10.5, from ruby30.

Also adds self as co-maintainer.

TESTED:
Built successfully with no variants on 10.4-10.5 ppc, 10.4-10.6 i386,
10.5-12.x x86_64, and 11.x-14.x arm64.
Built with all variants except jemalloc and yjit on 10.4-10.5, all
variants except yjit on 10.6-10.13, and all variants on 10.14+.
Tests not run due to test framework issues.