ntpsec: Fix livecheck
The previous regex change to allow the patch letter broke the case
where the letter is absent.  The regex is now derived from the GitHub
default regex, but adjusted to exclude patterns with dashes to avoid
versions such as X.Y.Z-rcN.  The unnecessary quote exclusion is
dropped.

TESTED:
Tested with current and older port versions, including with modified
regex to make 1.2.2a the apparent latest.