asciidoctor: update to v2.0.22
Also updates the underlying ruby version to 3.3, just to avoid EOL
concerns (since it doesn't affect the behavior of asciidoctor itself).
3.3 is outside the compatible version range stated on the asciidoctor
homepage, but that appears to be outdated.

TESTED:
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-12.x x86_64, and
11.x-14.x arm64.
Builds and passes tests on all versions.
Functionally tested with gpsd and ntpsec on 10.9.  Generated docs
match except for expected metadata differences specifically related to
asciidoctor version.