polyml: Update to 5.9.1
Update to 5.9.1 and remove the upstream patches that it includes. Also
some portfile cleanup:

Add standard modeline.

Remove "platforms darwin"; that's the default.

Use https homepage since the http page redirects to it.

Use "github.tarball_from archive" instead of the deprecated default.

Use path:-based dependency for pkgconfig so pkgconf could satisfy it.

Remove --mandir configure arg since the manpages already go to the right
place without that.

Not sure if the --build arg is still needed but use configure.build_arch
for that instead of build_arch.

Add --without-x configure arg so that X11 is not used even if it is
installed.

Ignore implicit declaration of strchr for the usual reason.

Use leading 0 for octal numbers.