cmake-bootstrap: disable ncurses GUI on Tiger
the ncurses header installed by macports ncurses port is difficult
to avoid when building on Tiger, and there is an inconsistency with the
system ncurses library.

proper fixes are discussed in https://trac.macports.org/ticket/69990
but for the time being, disable to the (unneceessary) cmake ncurses GUI

see: https://trac.macports.org/ticket/69990

NB. I'm not revbumping this port for everyone just for this, nor am I
adding a special revbump for Tiger.