opam: add dependency on gpatch
Updating the opam package repository can silently fail
when using macos patch whenever files are removed upstream.
As of opam 2.1.6 there is a warning on install to
manually install gpatch to avoid any possible issue
and the corruption of the local switches. See also
https://github.com/ocaml/opam-repository/issues/25961

This commit adds a dependency to gpatch to prevent
any issue and avoid unnecessary manual intervention.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>