bun: update to version 1.1.12 and use Node.js 22 to build
As well as version number and checksum changes, this commit also
explicitly sets the npm.nodejs_version option to 22. This is because by
default the npm port group depends on two conflicting node ports (It
depends directly on node20, and on npm10, which depends on node22). This
can be removed in a future update if the node port group's defaults are
changed.