ruby33: update to v3.3.2
See:
https://www.ruby-lang.org/en/news/2024/05/30/ruby-3-3-2-released/

As usual, all source patches are in a single patchfile derived from
the ruby fork at github.com/fhgwright/ruby.  The changes are much more
readable in the git repo.

The patches are functionally unchanged since 3.3.1, though the patchfile
is updated to reflect the newer mtimes.

Also adds comments near the version to clarify the update procedure.

TESTED:
Built successfully with no variants on 10.4-10.5 ppc, 10.4-10.6 i386,
10.5-12.x x86_64, and 11.x-14.x arm64.
Built with all variants except jemalloc and yjit on 10.4-10.5, all
variants except yjit on 10.6-10.13, and all variants on 10.14+.
Tests not run due to test framework issues.