ruby33: Cosmetic patchfile cleanup
This regenerates the patchfile from updated git commits, and updates
the related comment in the Portfile.  There are no substantive
changes in the Patchfile, but the mtimes are updated to correspond
to the new release.  The result of the patch phase is unchanged.

Also, the update URL was omitted from the original update commit
message.  See:

https://www.ruby-lang.org/en/news/2024/06/12/ruby-3-3-3-released/

TESTED:
Verified that the build trees from "port patch" are identical before
and after this change.