libsrtp/libsrtp1: build shared library
the port previously built a static library, however
linking failed as the port uses openssl but the
openssl libraries are not specified in the standard
pkgconfig libs.

rather than modify every user of these libraries to
call for the pkgconfig flags with '-static', build
the shared library version of this software, that is what
is preferred in macports anyway.