rxvt: Fix build with newer compilers
The build system does not exit with an error if a build error occurs;
instead, the program that could not be built simply doesn't get
installed. This fixes build errors that occurred with newer compilers
that prevented the rxvt executable from building.

Closes: https://trac.macports.org/ticket/70310

Also fix the rclock Makefile so that it builds during the build phase
instead of waiting until the destroot phase.