ruby33: Cosmetic patchfile cleanup
This regenerates the main patchfile from updated git commits, removes
the additional patchfile, and updates the related comment in the
Portfile.  There are no substantive changes in the patchfile(s), but
the mtimes are updated to correspond to the new release.  The result
of the patch phase is unchanged.

Also, the update URL was omitted from the original update commit
message.  See:

https://www.ruby-lang.org/en/news/2024/07/09/ruby-3-3-4-released/

TESTED:
Verified that the build trees from "port patch" are identical before
and after this change.
Built successfully with no variants on 10.4-10.5 ppc, 10.4-10.6 i386,
10.5-12.x x86_64, and 11.x-14.x arm64.
Built with all variants except jemalloc and yjit on 10.4-10.5, all
variants except yjit on 10.6-10.11, and all variants on 10.12+.
Tests not run due to test framework issues.