clang-17,18: disable verbose_abort in availability
when availability tests are disabled (macports-libcxx, darwin 10)
a fallback set of defines are used in

include/c++/v1/__availability

however, the fallback defaults to indicating that libcxx contains
a verbose_abort function. This is currently always false for MacPorts
uses, and so link errors occur.

disable this by default, so a fallback pathway is used

see: https://trac.macports.org/ticket/68640

this has no effect on the normal functionality of this header,
as the __APPLE__ pathway is used otherwise.