py-meson: Update patch-meson64-32bit-apple.diff
"sudo port install glib2" was failing on OS X 10.4 Tiger (32-bit) with meson build reporting "subprocess" being an undefined symbol.
The above used to happen because patch-meson64-32bit-apple.diff introduces some code which depends on "subprocess" python module but the module itself is not imported.
This change updates patch-meson64-32bit-apple.diff file to include an import statement for the "subprocess" module to resolve the described failure.