vigor: Delete ancient unbuildable useless port
This software has not been updated since 2000 and has not been buildable
in MacPorts since we upgraded tcl to 8.6 in 2012. This software appears
to be a joke version of vi that includes an unhelpful assistant modeled
after one in another word processing program. I can't see any value in
having this in MacPorts.

Closes: https://trac.macports.org/ticket/61292
Closes: https://trac.macports.org/ticket/70153