gelide: Fix fetch failure
Fetching from a SourceForge git repository snapshot is not reliable so
go back to fetching the last released version and applying patches for
all the changes that were made in the repository after that.

Closes: https://trac.macports.org/ticket/57290

The developer has the bad habit of combining several unrelated changes
into a single commit. I've separated these out into one patchfile per
logical change. The developer's change descriptions were incomplete,
badly worded, or sometimes completely absent, so I wrote new ones.

One of the changes included a new image file. I included the image file
in the files directory and copy it into place in the Portfile.

I've also included a new patch to fix build failure with the version of
glibmm that we have in MacPorts now.

The revision is increased because a problem with displaying the names of
the translators in the about box has been fixed by omitting an erroneous
upstream change and the version of the GPL displayed in the about box
has been corrected.