ruby31: Remove inappropriate -std=c99.
The Portfile has been carrying the -std=c99 setting (for OS <10.7)
forward through many versions, but that has been inappropriate since
ruby30, when the compiler selection was switched by the thread-local
storage requirement.  Since the compiler now defaults to C11,
specifying C99 is a step backward.  This removes that inappropriate
setting.

Since this is not known not to impact the installed content, it
includes a revbump.

Also, since the latest legacy-support includes fgetattrlist(), builds
for <10.6 no longer include the private fallback, making the installed
content different on those platforms.  The revbump covers that as
well.

Also removes unnecessary "platforms darwin", making lint happy.

TESTED:
Minimally tested on unaffected 10.7+.
Built with all variants except jemalloc on 10.5 ppc, and all variants
on 10.6 i386 and 10.6 x86_64.
Builds on other affected platforms known to be broken (as before).
Tests not run due to test framework issues.