ruby32: Fix build with gcc14.
This backports the upstream commit 055613fd86, which fixes a
pointer-type mismatch in 32-bit builds, which gcc14 treats as an
error.

Although the intent is only to fix a build failure, the fix replaces
one function call with another, and hence requires a revbump.

TESTED:
Built successfully on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-12.x
x86_64, and 11.x-14.x arm64.  Included all variants compatible with
available dependencies on the respective platforms (except ppc builds,
which were limited to default variants to save time).