stylua: Enable more Lua version formatting
By default, stylua is built with support for Lua 5.1 only, but can be
built to support Lua 5.1 through Lua 5.4 and Luau. This only requires
adding `luau` and `lua54` to the default features, `editorconfig` and
`wasm-bindgen`. `lua54` implies `lua52` and `lua53`.

I was able to format a Lua 5.4 file that was not previously formattable
with the existing feature set.