ruby33: update to v3.3.5
See:
https://www.ruby-lang.org/en/news/2024/09/03/3-3-5-released/

As usual, all source patches are in a single patchfile derived from
the ruby fork at github.com/fhgwright/ruby.  The changes are much more
readable in the git repo.

The patches no longer include the atomic_signal_fence fix, which is
now upstream.  Otherwise, the only patchfile changes are for the newer
file mtimes.

Also updates the +yjit description to indicate that it requires rust.

TESTED:
Built successfully with no variants on 10.4-10.5 ppc, 10.4-10.6 i386,
10.5-12.x x86_64, and 11.x-14.x arm64.
Built with all variants except jemalloc and yjit on 10.4-10.5, all
variants except yjit on 10.6-10.12, and all variants on 10.13+.
Tests not run due to test framework issues.