alpine: Fix build with Xcode 16/clang 16 (macOS 15)
Alpine’s configure script contains a check to detect the parameter type
accepted by qsort’s comparison function. It does this by doing a trial
compilation of a program with a comparison function that accepts two
void* parameters, which is the modern form specified by the standard,
and checking whether compilation succeeded. If compilation fails, it
assumes the historic pre-standard form, which uses char* parameters, and
configures Alpine accordingly.

The program for the test comparison is written in an archaic style that,
among other things, does not explicitly declare the comparison
function’s return type. This can be the subject of a compiler warning,
-Wimplicit-int.

Clang 16, included in Xcode 16, enables -Werror=implicit-int by
default[1], as does GCC 14[2]. Because this condition is now an error,
the trial compilation began failing under these compilers, causing
Alpine to be configured to use char* parameters in qsort comparison
functions. Because this does not match qsort’s actual declaration, where
comparison functions are expected to use void* parameters, this would
result in a build failure.

The program used to detect the proper type to use for qsort’s comparison
functions is updated to avoid the -Wimplicit-int problem.

[1] https://releases.llvm.org/16.0.0/tools/clang/docs/ReleaseNotes.html#potentially-breaking-changes
[2] https://gcc.gnu.org/gcc-14/porting_to.html#warnings-as-errors