ruby30: Fix build with Xcode 16.
This backports an upstream fix that was already included in ruby3.1+.
Although it's unlikely that this changes the content of successful
builds, we revbump anyway out of conservatism.

TESTED:
Built successfully with default variants on OSX 10.4-10.5 ppc,
10.4-10.6 i386, 10.5-12.x x86_64, and 11.x-15.x arm64.  Also built
successfully with all -universal-jemalloc variants on 10.4-10.5.  Also
built successfully with all variants on all 10.6+ cases.
Tests not run due to test framework issues.