flang-14: fix for macOS ≥ 14
93df94134805 (as fixed by c04f7bf1e4b0) must be made to apply to
flang-14 in addition to clang-14, because the sanitizer runtimes are
built for both.

flang-mp-14 delegates to a bash wrapper script that uses modern
bash-isms not supported by bash-3.2 shipped by Apple in /bin. A run-time
dependency on the bash port is added, and the wrapper script is patched
to use MacPorts’ bash as its interpreter rather than whatever happens to
be first in the PATH.