openssh: Restore "platforms any" to ssh-copy-id subport
The "platforms any" line was inappropriately removed from the
'ssh-copy-id' subport by commit 11c25bbb3cd to get rid of a spurious
lint warning.  But "platforms any" is not the same as "platforms
darwin", is not the default, and may be useful for non-Mac installs.
The lint warning in this case is just a bug that should be ignored.

This only affects the 'ssh-copy-id' subport, and does not require a
revbump.

TESTED:
The spurious lint warning is back as expected, but both ports build
successfully on all platforms.