flang-{15–18}: depend on and use bash (#26049)
* flang-15: depend on and use bash

flang-to-external-fc-mp-15 delegates to a bash wrapper script that uses
modern bash-isms not supported by bash-3.2 shipped by Apple in /bin. A
run-time dependency on the bash port is added, and the wrapper script is
patched to use MacPorts’ bash as its interpreter rather than whatever
happens to be first in the PATH.

* flang-16: depend on and use bash

flang-to-external-fc-mp-16 delegates to a bash wrapper script that uses
modern bash-isms not supported by bash-3.2 shipped by Apple in /bin. A
run-time dependency on the bash port is added, and the wrapper script is
patched to use MacPorts’ bash as its interpreter rather than whatever
happens to be first in the PATH.

* flang-17: depend on and use bash

flang-to-external-fc-mp-17 delegates to a bash wrapper script that uses
modern bash-isms not supported by bash-3.2 shipped by Apple in /bin. A
run-time dependency on the bash port is added, and the wrapper script is
patched to use MacPorts’ bash as its interpreter rather than whatever
happens to be first in the PATH.

* flang-18: depend on and use bash

flang-to-external-fc-mp-18 delegates to a bash wrapper script that uses
modern bash-isms not supported by bash-3.2 shipped by Apple in /bin. A
run-time dependency on the bash port is added, and the wrapper script is
patched to use MacPorts’ bash as its interpreter rather than whatever
happens to be first in the PATH.