openjdk{17,21}: workaround for regression in JDK-8266242
Adds temporary patch to CGraphicsDevice.m to guard use of undeclared
architecture ARM64 unknown to macOS SDKs before macOS 11.

openjdk versions before 22 support build on 10.12 - 10.15, but when
backporting the fix for 22 to 17 and 21, this support was broken.
This patch should be removed once upstream regression is addressed.

See: https://trac.macports.org/ticket/71049