erlang: update to 27.1.2
Also fix support for JIT which was disabled by default by upstream on
macOS/amd64 because of an annoying popup starting from Sonoma. Fix was
eventually provided by asmjit and this is backported here as a patch.

Also fix issue with symlinks to erlang tools that was broken with erlang 27
upgrade.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>