aircrack-ng: Remove bad patch that broke PowerPC build
The patch was an attempted backport of an arm64 build fix from John the
Ripper: openwall/john@c9825e688d1fb9fdd8942ceb0a6b4457b0f9f9b4

But the patch was not backported properly, and instead of changing
anything for the arm64 build, it introduced a syntax error for the
PowerPC build.

The problem the patch was trying to fix was already fixed in aircrack-ng
1.7: aircrack-ng/aircrack-ng@a4cdb89cae06545d547a6c15a5a92f7972fad38d

Also indicate that C++11 is required and use path dep for pkgconfig.

Closes: https://trac.macports.org/ticket/71326