gcc9: use no-pie patch on systems < 10.6
building gcc9 on 10.5 Intel fails with the no-pie bug noted in the patch explanation. gcc9 builds with clang-11 on 10.5 Intel, so it's not surprising it hits this no-pie bug. 

the patch was restricted to 10.6+ when a year or two ago it was noted that the patch caused a build to fail on a different system setup.

this may need further tweaking as more systems try to build this, but for the moment, libgcc9 is needed on 10.5 Intel to move it to gcc14/libgcc14, and it won't build without the patch.