Gyoto: update to 2.0.2+20241208
Use a git commit as build system has been fixed since 2.0.2.

Remove dependency on Yorick as building with it is considered
deprecated and it breaks building Gyoto on certain architectures.

Remove "universal_variant no" as it was due to an incompatibility with
Yorick.

Remove build conflict on self as it doesn't seem to cause issues
anymore.