libvpx, libvpx-devel: Resolve self-conflict
We don't need or want to link with any MacPorts libraries so remove
-L/opt/local/lib from LDFLAGS which the build system inserts before -L.
and remove -I/opt/local/include from CPPFLAGS because although the build
doesn't use CPPFLAGS it could do so in a future version.

Fixes build failure when an older version of libvpx or one of a
different architecture is already installed.

Closes: https://trac.macports.org/ticket/66259

Fixes build failure when the gtest port is installed.

Closes: https://trac.macports.org/ticket/70989