ntpsec: update to 1.2.4
As of this version, legacy-support is used to provide clock_gettime()
and clock_settime(), though patches are still needed to accommodate
other missing features prior to 10.13.  As usual (and as documented in
the Portfile), the patches can be viewed more readably on the GitLab
fork.

The default Python variant is now python312.

Former versions opportunistically included mDNS support, relying on an
undeclared dependency on the avahi port.  This is now a variant, with
an appropriate dependency.  Since this is a fairly low-utility
feature, and avahi is a somewhat heavyweight port (that doesn't build
on all platforms), and that feature was never guaranteed to be
available, this variant defaults off.

TESTED:
Built (including running tests) on 10.4-10.5 ppc, 10.4-10.6 i386,
10.4-10.15 x86_64, and 11.x-15.x arm64.
Ran operationally on a representative subset.