grpc: Use compiler thread-local storage if available
but don't require it.

The upstream code contains a compatibility implementation of
thread-local storage using pthread which is used if the compiler doesn't
support TLS natively but the code failed to compile if the compiler
didn't support __has_feature.

Our patch to fix that inadvertently reversed the condition so that
compilers having TLS support would use the compatibility code and
compilers not having TLS support would fail to build. We subsequently
marked the port as requiring a compiler with TLS support.

Fix the condition so that compilers with TLS support will use it and
those that don't will use the compatibility implementation. Don't
require a TLS-supporting compiler anymore (although we do require a
C++17-supporting compiler, and maybe all of those also support TLS).

Closes: https://trac.macports.org/ticket/65525