libunwind/libunwind-headers: use on < darwin11
the functionality of the libunwind library was added
to libSystem in darwin11. Installing this port on any
system newer than darwin10 causes problems, and usually
causes an older libunwind to be used than the one already in
libSystem.

closes: https://trac.macports.org/ticket/66250

For interest:
a current libunwind is currently installed  as part of the clang builds, eg.:

% port contents clang-19 | grep libunwind
  /opt/local/libexec/llvm-19/include/__libunwind_config.h
  /opt/local/libexec/llvm-19/include/libunwind.h
  /opt/local/libexec/llvm-19/include/libunwind.modulemap
  /opt/local/libexec/llvm-19/lib/libunwind/libunwind.1.0.dylib
  /opt/local/libexec/llvm-19/lib/libunwind/libunwind.1.dylib
  /opt/local/libexec/llvm-19/lib/libunwind/libunwind.a
  /opt/local/libexec/llvm-19/lib/libunwind/libunwind.dylib

if testing a newer libunwind is desired on new systems, one can be found there.