ruby34: new port, version 3.4.1
See:
https://www.ruby-lang.org/en/news/2024/12/25/ruby-3-4-1-released/

The upstream code already incorporates the fixes from v3.2.6, but not
the fixes from 3.2.6_1, which are included in the patchfile here.

The legacy-support threshold is updated for clock_gettime() support.

The upstream linker warning suppression hackery was changed in v3.4 to
somthing badly broken.  This reinstates the 3.3 version, which is
needed for builds to work with certain library mismatches.

Ruby 3.4 does not build with Xcode 10.1 clang 425 on 10.7, so this
blacklists clangs < 500 (in addition to the preexisting blacklisting
of clang 900).

Also cleans up a couple of long-standing Portfile issues.

TESTED:
Built successfully on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-12.x
x86_64, and 11.x-15.x arm64.  Included all variants compatible with
available dependencies on the respective platforms.