py-gssapi: Mention GSSAPI source in longdesc
This change updates the port's long_description to specify which GSSAPI
implementation it is using.

py-gssapi needs a GSSAPI/Kerberos library to build against.  On Mac OS X
10.8 and earlier, py-gssapi uses the `kerberos5` port.  On Mac OS X 10.9
and later, py-gssapi uses the OS-provided GSS Framework.

The specific GSSAPI implementation is useful to know, because a Kerberos
credential that is kinit-ed from MIT Kerberos might not be visible to a
program that is using macOS' GSS Framework.