zmq: Re-enable CURVE encryption
zeromq 4.3.5 changed the default for the ENABLE_CURVE option to false
instead of true while removing the tweetnacl implementation because of
an inability to license (see
https://github.com/zeromq/libzmq/pull/4554).  The macports 4.3.5 version
bump, however, didn't specify the option and so currently doesn't
support secure zmq connections at all, while 4.3.4 and earlier *always*
supported CURVE regardless of the variant (either via libsodium or the
build-in tweetnacl, depending on whether the sodium variant was used).

This reenables it behind a new, default-enabled `curve` variant (which
depends on the already-default `libsodium` as that is now the only
supported zeromq curve implementation) so that the default variant zmq
package supports curve encryption, as it did for versions before 4.3.5.

I kept it distinct from the sodium variant because zmq technically can
make use of libsodium even when curve is disabled.