libffi: Fix build on Sequoia.
This cherry-picks an upstream fix for a build problem with a recent
clang.  Although the fix is included in v3.4.7, the latter has
significantly more test failures than v3.4.6.  Until that is
investigated further, we stay on v3.4.6 with the cherry-picked fix, to
avoid a possible functional regression.

Due to the (admittedly unlikely) possibility that this changes the
installed content, it includes a revbump, which is not very expensive
for this particular port.  No revbumping of dependents is needed.

Closes: https://trac.macports.org/ticket/72271

The unified patchfile is also renamed with a more generic name.

TESTED:
Successfuly built -/+universal on 10.4-10.5 ppc, 10.5-10.6 ppc (i386
Rosetta), 10.4-10.6 i386, 10.4-12.x x86_64, and 11.x-15.x arm64.
Tests not included due to many known failures.