python313: fix build on Tiger and Leopard
This adds the 2 patches from kencu@, originally posted in
https://trac.macports.org/ticket/71206#comment:13, to fix
the build on Tiger.

To fix the build on Leopard, we simply add
`-D_DARWIN_USE_64_BIT_INODE`.

Tiger does not have 64-bit inode support. Disabling the
assert completely is not ideal but for now is better
than not being able to build it at all.