mercurial: unbreak build
Mercurial seems to be in the process of moving from using a
Makefile-based build to a more normal Python-style build -- the Makefile
is still capable of building a wheel, but does so using Pip in a `venv`.
The previous solution was to try to call the GNU Make as a separate
invocation post-build, but that lost all sorts of environment variables.

As a temporary workaround, simply drop the parts that _must_ be built
with Make:

* man-pages
* /opt/local/bin/chg
* /opt/local/bin/rhg (+rust only)

This should allow the port to build, while retaining the most important
functionality while I try to fix it.

See: https://trac.macports.org/ticket/72455