mercurial: really unbreak build
The precise order of different `post-destroot` doesn't seem
well-defined, as sometimes, the one in the port gets executed before the
symlinking done by Python PortGroup. Regardless, this inconsistency
exposed a bug: We should of course be fixing the shebang in the *target*
if the symlink in `/opt/local/bin/hg`.

See: https://trac.macports.org/ticket/72455